+++ BE/issue_tracker/src/main/java/team02/issue_tracker/dto/issue/AbstractIssueResponse.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 import team02.issue_tracker.domain.Issue;
 import team02.issue_tracker.domain.IssueAssignee;
 import team02.issue_tracker.domain.IssueLabel;
-import team02.issue_tracker.dto.IssueLabelResponse;
 import team02.issue_tracker.dto.UserResponse;
 
 import java.time.LocalDateTime;
