+++ BE/issue_tracker/src/main/java/team02/issue_tracker/dto/issue/DetailIssueMilestoneResponse.java
@@ -1,4 +1,4 @@
-package team02.issue_tracker.dto;
+package team02.issue_tracker.dto.issue;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,14 +6,14 @@ import team02.issue_tracker.domain.Milestone;
 
 @AllArgsConstructor
 @Getter
-public class MilestoneDetailResponse {
+public class DetailIssueMilestoneResponse {
 
     private Long id;
     private String title;
     private int totalIssues;
     private int openIssues;
 
-    public MilestoneDetailResponse(Milestone milestone) {
+    public DetailIssueMilestoneResponse(Milestone milestone) {
         this(milestone.getId(), milestone.getTitle(), milestone.getTotalIssueCount(), milestone.getOpenIssueCount());
     }
 }
