+++ BE/issue_tracker/src/main/java/team02/issue_tracker/dto/issue/DetailIssueResponse.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 import team02.issue_tracker.domain.Issue;
 import team02.issue_tracker.domain.Milestone;
 import team02.issue_tracker.dto.CommentResponse;
-import team02.issue_tracker.dto.MilestoneDetailResponse;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,7 +12,7 @@ import java.util.stream.Collectors;
 @Getter
 public class DetailIssueResponse extends AbstractIssueResponse {
 
-    private Optional<MilestoneDetailResponse> milestone;
+    private Optional<DetailIssueMilestoneResponse> milestone;
     private List<CommentResponse> comments;
 
     public DetailIssueResponse(Issue issue) {
@@ -25,10 +24,10 @@ public class DetailIssueResponse extends AbstractIssueResponse {
                 .collect(Collectors.toList());
     }
 
-    private Optional<MilestoneDetailResponse> toMilestoneDetailResponse(Milestone milestone) {
+    private Optional<DetailIssueMilestoneResponse> toMilestoneDetailResponse(Milestone milestone) {
         if (milestone == null) {
             return Optional.empty();
         }
-        return Optional.of(new MilestoneDetailResponse(milestone));
+        return Optional.of(new DetailIssueMilestoneResponse(milestone));
     }
 }
